backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.9.1
python-dateutil==2.9.0.post0